Avoid reusing old session ID on when starting the device
When I turn off the device and turn it on the next day, I think it tries
to resuse the old session ID, which adds a 5 minute delay before seeing
the bgs.

On startup check the last session response time. If it's been more than
one hour, start from scratch.

I haven't tested this yet. If I don't revert this tomorrow that means it
worked.